In Thaler v. Perlmutter (2023), the Federal District Court for D.C. ruled that “Underlying that adaptability, however, has been a consistent understanding that human creativity is the sine qua non at the core of copyrightability”. 17 U.S.C. § 102(a) says that “Copyright protection subsists [..] in original works […] either directly or with the aid of a machine or device”.
My question is outside of the scope of this lawsuit: do prompts to AI count as a human using the “aid of a machine or device” to create a creative work? Or, is the transformation from a simple textual prompt to a graphical representation considered transformative under Campbell and 17 U.S.C. § 107, such that the AI is the creator of the secondary graphical work, to the extent that it is not a derivative work of the text prompt? Or would the prompt simply be considered an idea, which is not copyrightable under Baker v. Selden?